+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added: The `-C` option can be used to set the working directory for `git-branchless` commands.
 - Added: The `--hidden` option can be passed to `git smartlog` to show commits which are not ordinarily visible.
+- Changed: Git configuration is written to a file under `.git/branchless`, instead of writing it directly to `.git/config` (which may clobber user settings).
 - Fixed: Output of subcommands is no longer overwritten by progress updates.
 - Fixed: Improved performance up to 100x for commit deduplication during  `git move` when rebasing past certain large commits.
 - Fixed: Improved performance up to 10x for smartlog rendering.